src/requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiohttp-retry==2.5.1
aiosignal==1.2.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
atpublic==2.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.8.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
commonmark==0.9.1
configobj==5.0.6
contextvars==2.4
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.8.0
dpath==2.0.6
dulwich==0.20.46
dvc==2.8.1
entrypoints==0.4
et-xmlfile==1.1.0
facenet-pytorch==2.5.2
filelock==3.4.1
flatten-dict==0.4.2
flufl.lock==3.2
frozenlist==1.2.0
fsspec==2022.1.0
ftfy==6.0.3
funcy==1.17
future==0.18.2
gdown==4.5.1
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.13.0
google-auth-oauthlib==0.4.6
grandalf==0.6
grpcio==1.48.2
idna==3.4
idna-ssl==1.1.0
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.2
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.1
kiwisolver==1.3.1
mailchecker==5.0.3
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
multidict==5.2.0
mypy-extensions==0.4.3
nanotime==0.5.2
natsort==8.2.0
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.6
networkx==2.5.1
notebook==6.4.10
numpy==1.19.5
oauthlib==3.2.2
openpyxl==3.0.10
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pathspec==0.8.1
phonenumbers==8.12.57
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
pygit2==1.6.1
Pygments==2.13.0
pygtrie==2.5.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-benedict==0.26.0
python-dateutil==2.8.2
python-fsutil==0.7.0
python-slugify==6.1.2
pytz==2022.4
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
shortuuid==1.0.9
shtab==1.5.5
six==1.16.0
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.2.post1
tabulate==0.8.10
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.12.1
testpath==0.6.0
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
tomli==1.2.3
torch==1.10.2
torchvision==0.11.2
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typed-ast==1.5.4
typing_extensions==4.1.1
urllib3==1.26.12
voluptuous==0.13.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.6.1
xlrd==2.0.1
xmltodict==0.13.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.6.0
